# PERF: datetime parsing with Q