```
In [3]: %timeit pd.Timestamp("2014Q1")
3.04 µs ± 63.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  # <- main
2.8 µs ± 19.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  # <- PR
```